moved browser addition to phonegap module
