+++ lib/cli/serve.js
@@ -35,44 +35,6 @@ module.exports = function(argv, callback) {
         refresh: argv.refresh,
         verbose: argv.verbose
     };
-    var startWatchandServe = function() {
-            startWatch(data, function() {});
-        };
-
-    if(argv.browser){
-
-        addBrowser(startWatchandServe);
-    } else {
-        startWatch( data, callback);
-    }
-    
-
-    // gracefully exit on ctrl-c
-    process.on('SIGINT', function() {
-        process.exit();
-    });
-};
-
-function startWatch(data, callback) {
-    // add gaze instance to data(options)
-    var watches = [path.join(process.cwd(), 'www/**/*')];
-    var watch = new gaze.Gaze(watches);
-    data.watch = watch;
-
-    //only do cordova prepare
-    watch.on('all', function(filepath) {
-        var options = {};
-        options.cmd = 'cordova prepare';
-        pg.cordova(options, function() {
-        });
-    });
-
-    watch.on('ready', function(watcher){
-        startServing(data, function(){});
-    })
-}
-
-function startServing(data, callback) { 
     phonegap.serve(data, function(e, server) {
         if (!e) {
             console.log('');
@@ -81,20 +43,8 @@ function startServing(data, callback) {
         }
         callback(e);
     });
-}
-
-function addBrowser(callback) {
-
-    if (!fs.existsSync(path.join(process.cwd(), 'platforms/browser'))) {
-        var options = {};
-        options.cmd = 'cordova platform add browser';
-        pg.cordova(options, callback);
-
-    } else {
-         callback();
-    }
-
-}
-
-module.exports.addBrowser = addBrowser;
-
+    // gracefully exit on ctrl-c
+    process.on('SIGINT', function() {
+        process.exit();
+    });
+};
