+++ lib/phonegap/serve.js
@@ -6,10 +6,12 @@ var Command         = require('./util/command'),
     project         = require('./util/project'),
     server          = require('connect-phonegap'),
     proxy           = require('./util/connect-proxy'),
-    util            = require('util');
+    fs              = require('fs'),
+    path            = require('path'),
+    util            = require('util'),
+    gaze            = require('gaze');
 
-
-/**
+/**cd testSer
  * Server Default Settings
  */
 var ServeDefaults = {
@@ -98,18 +100,34 @@ AppCommand.prototype.execute = function(options, callback) {
     // change to project directory and delegate errors
     if (!project.cd({ emitter: self.phonegap, callback: callback })) return;
 
+    /**
+     * Start Server
+     */
+    var startWatchandServe = function() {
+        startWatch(options, callback ,self);
+    };
+
+    if(options.browser) {
+        addBrowser(startWatchandServe,self);
+    } else {
+        startWatch( options, callback, self);
+    }
+};
+
+function startWatch(data, callback, self) {
+    // add gaze instance to data(options)
+    var watches = [path.join(process.cwd(), 'www/**/*')];
+    var watch = new gaze.Gaze(watches);
+    data.watch = watch;
     var _errorHandler = function(err) {
         self.phonegap.emit('error', err);
         callback(err);
     };
 
-    /**
-     * Start Server
-     */
-    var startServer = function (self, options) {
+    watch.on('ready', function(watcher) {
         self.phonegap.emit('log', 'starting app server...');
 
-        server.listen(options)
+        server.listen(data)
               .on('browserAdded', function() {
                   self.phonegap.emit('browserAdded');
               })
@@ -121,19 +139,24 @@ AppCommand.prototype.execute = function(options, callback) {
                   self.phonegap.emit('log', statusCode, url);
               })
               .on('update', function(c) {
-                  if(options.connect === true) proxy.uploadUpdatedZip(options);
+                  if(data.connect === true) proxy.uploadUpdatedZip(data);
               })
               .on('complete', function(data) {
                   callback(null, data);
-                  if(options.connect === true) proxy.uploadAppZip(options);
+                  if(data.connect === true) proxy.uploadAppZip(data);
               });
-    };
-
-
-    var cordovaCommand = 'cordova prepare';
-
-    // use cordova to prepare the project from the command-line
-    self.phonegap.cordova({ cmd: cordovaCommand }, function(e) {
-        startServer(self, options);
     });
-};
+}
+
+
+function addBrowser(callback,self) {
+    if (!fs.existsSync(path.join(process.cwd(), 'platforms/browser'))) {
+        var cordovaCommand = 'cordova platform add browser';
+        // use cordova to prepare the project from the command-line
+        self.phonegap.cordova({ cmd: cordovaCommand }, callback);
+
+    }
+    else {
+        callback();
+    }
+}
